% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\docType{data}
\name{data_intervals}
\alias{data_intervals}
\title{Intervals Data for the Yahoo Finance API}
\format{
A data frame.
}
\usage{
data_intervals
}
\description{
A data frame with the available intervals data for the Yahoo Finance API.
}
\keyword{datasets}
