% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{get_col}
\alias{get_col}
\title{Get a Column from the Yahoo Finance API}
\usage{
get_col(data, col)
}
\arguments{
\item{data}{data frame or list. Data created using the \code{\link{get_data}}
function that contains an \code{index} column and the requested column.}

\item{col}{string. Column name to get (i.e., "open", "high", "low", "close", "adjclose", "volume").}
}
\value{
A data frame with rows as the \code{index} and columns as the symbols.
}
\description{
A function to get a column from data retrieved from the Yahoo Finance API.
}
\examples{
\dontrun{
data <- get_data(c("AAPL", "MSFT"))

adj <- get_col(data, "adjclose")
}
}
